Mark test_target_without_psdriver as an expected failure, improve error handling
There's currently no easy way to disable the webview for test_target_without_psdriver,
so mark the test as an expected failure for now.
It has been tested to pass on a freshly booted target at login screen (ie. without a webview created at that point).

Also added PSTargetException handling in other tests to mark them as skipped rather than failed
if the test target is unreachable for any reason.