Rework test_pstarget.py to ensure disconnect() is always called after a test
If any assertion fails, we exit the function without reaching the
disconnect() call, which makes the pytest test runner hang indefinitely
due to having a dualshock keythread still hanging in the background.

Move the PSTarget initialization and cleanup to setUp() and tearDown(),
as it should be anyway. Also split the PSTestCase class in two, one
for tests that need automatic target handling, and one were we manage
that as part of the test.