Get rid of the main cause of hangs in psdriver by disabling console log monitoring
The amount of log data sent to psdriver upon startup is what causes the hangs.
Note that this doesn't fix the bug in psdriver, but it fixes the hanging issue
without a need to change the psdriver code.

This feature can be re-enabled once the bug in psdriver has been fixed.