Hooked deci.Info and deci.Power functions to PSTarget
Renamed PowerStatus to PowerState in the process, and adjusted the enum name.
Unlike Dualshock and Console, added the functions directly under the PSTarget
API instead of hooking the objects, since their amount is limited and they
belong in PSTarget directly.

Note that both underlying protocol wrappers are using lazy instantiation,
to limit the amount of connections to create upfront.

Also added unit tests, even though the power ones are currently incomplete
due to limitations in the deci layer.