Introduce the Mantis plugin for automatically distributing tests among devkits
It is based on the xdist plugin, with customizations for our use case.
This was needed as the hooks provided weren't enough to customize the
behavior to our requirements.

Note that we're still using most of the xdist infrastructure, we're just
customizing the way nodes are initialized and in the future replacing the
LoadScheduler with our own.