Simplified DeciQueue code and added missing lock for the exception object
Removed a lot of fragile code, mainly from the fact that we had non-blocking
locks in the rwthread, which means we couldn't use them with 'with' and had
to handle the .release() manually. These don't need to be non-blocking, as
there's no chance of deadlock (the locks are never interleaved).
Making this change allows simplifying the code greatly, and removes the
need for _abort(). All locks will be properly released when the
with-blocks are exited, and we can now just save the exception directly in
there.

Added a required lock for the self._exception object, as it's queried on
both sides concurrently.

Also added missing stop() to clean up the thread when an exception is reported
in there. (we would exit the function, but thread.join() would never get
called)

Finally, adjusted some of the locks that were acquired longer than
necessary. (for example when calling select())

All tests in tests/deci and tests/pstarget pass, and I've actually noticed
a slight increase in performance when saving screenshots (which got
significantly slower through one of the past refactors).