Rework the logic for the psdriver server startup/shutdown and external servers
Now every process starts its own psdriver server on the first available
port >= 9515, incrementing the port number until an available port is
found.
Existing processes are not assumed to be external servers anymore.
Instead, add a new flag on PSDriverServer to indicate usage of an external
server instead of the defualt local server. This flag will be hooked up to
a cmdline argument in Mantis.

Also add a new psdriver_server session fixture and attach it to all pstarget
fixtures by default. This allows shutting down the server automatically at
the end of a test session.